+++ libs/common/array.go
@@ -1,5 +1,0 @@
-package common
-
-func Arr(items ...interface{}) []interface{} {
-	return items
-}