+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.22.1
+
+*July 5th, 2018*
+
+IMPROVEMENTS
+
+* Cleanup post repo-merge.
+* [docs] Various improvements.
+
+BUG FIXES
+
+* [state] Return error when EndBlock returns a 0-power validator that isn't
+  already in the validator set.
+* [consensus] Shut down WAL properly.
+
 ## 0.22.0
 
 *July 2nd, 2018*